tests: correct one qwery unit test
According to the documentation, `CSSselect()` should not query element itself except if an array of elements is used as context. This behavior differs from the one of Qwery library, so the unit test has to be changed.

See this comment for more information: https://github.com/cheeriojs/cheerio/issues/728#issuecomment-158897030