Fix regression, include self while querying array
Elements from an array are now included to the query.

The issue was due to `absolutize()` introduced in v1.1.0. This function adds `":scope "` to the selector, and because of the descendant selector `" "`, "root" element from the array could not be reached.

The fix is to make the descendant selector more flexible so it accepts current element (without call `getParent()`) if the context is an array, and if the descendant selector is directly following the first `:scope`. So, others descendant selectors work as usual.