Simplify document nodes for equality checking.
Mocha's object printer was getting stuck in an infinite loop, This also
prevents deepEqual from doing unnecessary equality checks but it's not
clear that was more expensive than the clone operation required by
simplification. The right fix for this is in the mocha
library. I've filed a [bug](https://github.com/mochajs/mocha/issues/3229);
When it's fixed, this patch can be reverted.