feat(adapter): Use `adapter.equals` across the codebase
We were very inconsistent in usage of `adapter.equals` before. With this change, we will always use it for equality checks.

BREAKING: The signature for pseudos has changed. The second argument is now the `options` object, not the adapter anymore. To update your code, extract the adapter from the options object, eg. `pseudos.custom = (elem, adapter) => {}` is now `pseudos.custom = (elem, {adapter}) => {}`.